+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 - Product: If purchase module is not installed, hide boolean "Define the shipping coef by partner" and "Shipping Coef."
 - Convert Lead: Removed separate panel of prospect and added prospect boolean into partner tab. Removed wrong field from opportunity tab. 
 - Convert Lead: Removed opportunity and events conversion. Just partner and contact created from lead with events and opportunity linked.   
+- Current user password should be required when updating user password 
+- Convert Lead: Fix translation and default values. 
 
 ## Bug Fixes
 - invoice: fix hilighting for overdue bills. Warning hilite was always taking precedence over danger because of overlapping conditions.
@@ -33,6 +35,7 @@
 - Fix empty partner list from batch report.
 - Convert Lead: Fix type,call type and lead field of converted events. 
 - Advanced Export: Fix export every fields of model.
+- Advanced export : Fix input issue of selection field for export.
 
 ## [5.0.0-rc1] - 2018-04-16
 ## Features